Rename component files
