+++ src/lib/sections/testnet.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { _ } from 'svelte-i18n';
 
-	import LinkButton from '$lib/components/LinkButton.svelte';
+	import LinkButton from '$lib/components/link-button.svelte';
 	import { testNetLink } from '../../consts';
 	import * as images from '../../consts/images';
 </script>
@@ -10,7 +10,9 @@
 	<section class="gap-4 md:pt-[16.875rem] pt-40 text-base-100 md:pb-[13.1215rem] pb-[7.5rem]">
 		<div class="flex gap-4 justify-center">
 			<img class="md:w-10 w-[1.65rem]" src={images.color} alt="logo" />
-			<span class="uppercase tracking-wider leading-none self-center md:text-xl text-sm">sunrise</span>
+			<span class="uppercase tracking-wider leading-none self-center md:text-xl text-sm"
+				>sunrise</span
+			>
 		</div>
 		<h1
 			class="text-center text-[2rem] leading-10 font-semibold md:text-6xl xl:text-[6rem] xl:leading-[7.375rem] md:leading-[6rem] tracking-wide"
