+++ src/lib/sections/token-sale/core-team.svelte
