+++ src/lib/sections/token-sale/sales-details.svelte
